+++ compiler-core/src/error/tests.rs
@@ -1,4 +1,5 @@
 use super::*;
+use insta::assert_debug_snapshot;
 use insta::assert_snapshot;
 
 #[test]
@@ -34,4 +35,18 @@ fn test_shell_program_not_found_error() {
             }
         }
     }
+}
+
+#[test]
+fn hex_session_revoked() {
+    let err = Error::HexSessionRevoked.to_diagnostics();
+    assert_debug_snapshot!(err[0]);
+}
+
+#[test]
+fn hex_error_conversion() {
+    assert_eq!(
+        Error::hex(hexpm::ApiError::OAuthRefreshTokenRejected),
+        Error::HexSessionRevoked
+    );
 }