+++ compiler-cli/src/config.rs
@@ -117,6 +117,7 @@ mod tests {
             source: ManifestPackageSource::Git {
                 repo: "repo".into(),
                 commit: "commit".into(),
+                path: None,
             },
         };
         assert_eq!(