+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 ### Build tool
 
+- The `gleam dev` command now accepts the `--no-print-progress` flag. When this
+  flag is passed, no progress information is printed.
+  ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
+
 ### Language server
 
 ### Formatter