+++ CHANGELOG.md
@@ -15,3 +15,7 @@
 ### Formatter
 
 ### Bug fixes
+
+- Fixed a bug where the build tool would check for new major versions of a local
+  or git dependency on Hex when running `gleam update`.
+  ([Giacomo Cavalieri](https://github.com/giacomocavalieri))